requirements.txt
discord.py
requests
opencv-python-headless
groq
gspread
google-auth
pytz
